Moving deprecated mbed-ls functions into detect module
